Objectlabs leafnode issue 1

+++ test/util.js
@@ -1,29 +1,45 @@
 var _ = require('lodash')
+var mongodbUri = require('mongodb-uri')
 
+var _o = require('@carbon-io/bond')._o(module)
 var oo = require('@carbon-io/atom').oo(module)
 var testtube = require('@carbon-io/test-tube')
 
 var connect = require('../lib/leafnode').connect
 
+// if we have a specific address in the environment for a stand alone instance, then use that,
+// otherwise default to localhost:27017
+STAND_ALONE_DB_URI = _o('env:LEAFNODE_STAND_ALONE_MONGODB_URI') || 'mongodb://localhost:27017'
+// if we have a replica set address in the environment, then initialize REPL_SET_DB_URI to that
+// NOTE: it is suggested that both be present, otherwise some connection tests will be skipped
+REPL_SET_DB_URI = _o('env:LEAFNODE_REPL_SET_MONGODB_URI') || undefined
+// if REPL_SET_DB_URI is present, use that for all tests, otherwise, use the stand alone instance
+DB_URI = REPL_SET_DB_URI || STAND_ALONE_DB_URI
+DB_NAME = mongodbUri.parse(DB_URI).database || 'leafnode'
+
 var LeafnodeTestSuite = oo({
   _type: testtube.Test,
   _C: function() {
-    this.dbUri = 'mongodb://localhost:27017'
+    this.dbUri = DB_URI
 
+    this.conOptions = undefined
     this.colName = undefined
+    this.con = undefined
     this.db = undefined
     this.c = undefined
   },
   setup: function() {
-    if (_.isUndefined(this.colName)) {
-      throw new Error('"colName" undefined')
+    this.con = connect(this.dbUri, this.conOptions)
+    this.db = this.con.db(DB_NAME)
+    if (!_.isUndefined(this.colName)) {
+      this.c = this.db.getCollection(this.colName)
     }
-    this.db = connect(this.dbUri)
-    this.c = this.db.getCollection(this.colName)
   },
   teardown: function() {
     try {
-      this.c.drop()
+      if (!_.isUndefined(this.c)) {
+        this.c.drop()
+      }
     } catch (e) {
       // ignore
     }
@@ -47,10 +63,13 @@ var LeafnodeTestSuite = oo({
 var LeafnodeTest = oo({
   _type: testtube.Test,
   setup: function() {
+    this.con = this.parent.con
     this.db = this.parent.db
     this.c = this.parent.c
     try {
-      this.c.drop()
+      if (!_.isUndefined(this.c)) {
+        this.c.drop()
+      }
     } catch (e) {
       // ignore
     }
@@ -61,5 +80,9 @@ var LeafnodeTest = oo({
 
 module.exports = {
   LeafnodeTestSuite: LeafnodeTestSuite,
-  LeafnodeTest: LeafnodeTest
+  LeafnodeTest: LeafnodeTest,
+  DB_URI: DB_URI,
+  DB_NAME: DB_NAME,
+  STAND_ALONE_DB_URI: STAND_ALONE_DB_URI,
+  REPL_SET_DB_URI: REPL_SET_DB_URI
 }
